ebay-db/scripts/requirements.txt
requests==2.31.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-api-python-client==2.122.0
